+++ src/OpenTelemetry.Exporter.OneCollector/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 0.1.0-alpha.2
+
+Released 2023-Mar-6
+
 * Update OpenTelemetry to 1.4.0
   ([#1038](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1038))
 
@@ -13,6 +17,9 @@
 * Switched to using a connection string design instead of passing
   instrumentation key directly.
   ([#1037](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1037))
+
+* Added `RegisterPayloadTransmittedCallback` API on `OneCollectorExporter<T>`.
+  ([#1058](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1058))
 
 ## 0.1.0-alpha.1